# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MAX.Локальный Туризм API",
    "description": "Backend-MVP персонального ИИ-планировщика поездок выходного дня: ИИ, Яндекс Расписания, geopy/Nominatim и погодные API.",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "MAX"
        ],
        "summary": "Auth Me",
        "operationId": "auth_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    },
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Profile",
        "operationId": "profile_api_v1_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    },
    "/api/v1/profile/onboarding": {
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Complete Onboarding",
        "operationId": "complete_onboarding_api_v1_profile_onboarding_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    },
    "/api/v1/cities/validate": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Validate City",
        "operationId": "validate_city_api_v1_cities_validate_get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 160,
              "description": "Город отправления из origins в GET /api/v1/app-config.",
              "title": "Origin"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 160,
              "title": "Destination"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Validate City Api V1 Cities Validate Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trips/generate": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Generate Trip",
        "operationId": "generate_trip_api_v1_trips_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    },
    "/api/v1/trips/jobs": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip Job",
        "operationId": "create_trip_job_api_v1_trips_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripJobCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    },
    "/api/v1/trips/jobs/{job_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip Job",
        "operationId": "get_trip_job_api_v1_trips_jobs__job_id__get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Trips",
        "operationId": "list_trips_api_v1_trips_get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip",
        "operationId": "get_trip_api_v1_trips__trip_id__get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/packing": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update Packing",
        "operationId": "update_packing_api_v1_trips__trip_id__packing_patch",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/reminders": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Reminders",
        "operationId": "get_reminders_api_v1_trips__trip_id__reminders_get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/reminders/{direction}": {
      "put": {
        "tags": [
          "trips"
        ],
        "summary": "Set Reminder",
        "operationId": "set_reminder_api_v1_trips__trip_id__reminders__direction__put",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "direction",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "outbound",
                "return_trip"
              ],
              "type": "string",
              "title": "Direction"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/share": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Share Trip",
        "operationId": "share_trip_api_v1_trips__trip_id__share_get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Share Trip Api V1 Trips  Trip Id  Share Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Publish Trip Share",
        "operationId": "publish_trip_share_api_v1_trips__trip_id__share_post",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Publish Trip Share Api V1 Trips  Trip Id  Share Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shared-trips/{token}/import": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Import Shared Trip",
        "operationId": "import_shared_trip_api_v1_shared_trips__token__import_post",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/geocode": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Geocode",
        "operationId": "geocode_api_v1_geocode_get",
        "security": [
          {
            "MaxInitData": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 160,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoPoint"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/destinations/suggest": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Suggest Destination",
        "operationId": "suggest_destination_api_v1_destinations_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationSuggestion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "MaxInitData": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Identity": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "max",
              "local"
            ],
            "title": "Mode"
          },
          "user": {
            "$ref": "#/components/schemas/MaxUser"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "user"
        ],
        "title": "Identity"
      },
...
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobState"
      },
...
      "TripJobCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobState",
            "default": "queued"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status_url"
        ],
        "title": "TripJobCreated"
      },
...
      "UserProfile": {
        "properties": {
          "onboarding_completed": {
            "type": "boolean",
            "title": "Onboarding Completed"
          },
          "preferences": {
            "$ref": "#/components/schemas/ProfilePreferences"
          }
        },
        "type": "object",
        "required": [
          "onboarding_completed"
        ],
        "title": "UserProfile"
      },
...
    },
    "securitySchemes": {
      "MaxInitData": {
        "type": "apiKey",
        "description": "Исходная window.WebApp.initData; подпись и срок действия проверяются сервером.",
        "in": "header",
        "name": "X-Max-Init-Data"
      }
    }
  }
}